oomd: fix potential race when killing cgroups
This came up when trying to enable the systemd-oomd extended test in
QEMU for Debian. In the case with one CPU, systemd-oomd would kill all
of the processes in a cgroup, pid1 would clean up that cgroup, and
systemd-oomd would get ENOENT trying to iterate the cgroup a final time
to ensure it was empty. To counter this, check explicitly for ENOENT.
Whether another candidate gets selected again or not will depend on
whether systemd-oomd sent SIGKILLs to any processes before it got the
ENOENT.

Fixes: 22029